LPD-56968 test: move all tests related to root model to one test class
